# out2/src/main/openapi/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: alberto.est.jimenez@gmail.com
    name: Alberto Jimenez
  description: Documentation for the client-accessible endpoints of the StackV orchestrator.
  title: StackV Orchestrator Developer API
  version: 0.8.2
servers:
- url: https://stackv.dev3.virnao.com/StackV-web/restapi/
...
paths:
...
  /meta:
    get:
      description: Retrieves all accessible metadata records.
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/paths/~1meta~1%7Bdomain%7D~1%7Bname%7D/post/responses/200/content/application~1json/schema"
                type: array
          description: List of SENSE Metadata objects.
        "500":
          $ref: "#/components/responses/Error"
      summary: Get All Records
      tags:
      - metadata
      x-accepts:
      - application/json
      x-tags:
      - tag: metadata
  /meta/{domain}:
    get:
      description: Retrieves all accessible metadata records under a specific domain.
      parameters:
      - description: "The domain of the metadata record, representing which platform\
          \ or third party it relates to."
        explode: false
        in: path
        name: domain
        required: true
        schema:
          example: JANUS
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/paths/~1meta~1%7Bdomain%7D~1%7Bname%7D/post/responses/200/content/application~1json/schema"
                type: array
          description: List of SENSE Metadata objects.
        "500":
          $ref: "#/components/responses/Error"
      summary: Get Domain Records
      tags:
      - metadata
      x-accepts:
      - application/json
      x-tags:
      - tag: metadata
  /meta/{domain}/{name}:
    delete:
      description: Removes a metadata record from the service.
      parameters:
      - description: "The domain of the metadata record, representing which platform\
          \ or third party it relates to."
        explode: false
        in: path
        name: domain
        required: true
        schema:
          example: JANUS
          type: string
        style: simple
      - description: The registered name of the metadata record.
        explode: false
        in: path
        name: name
        required: true
        schema:
          example: overview
          type: string
        style: simple
      responses:
        "204":
          description: An empty response.
        "500":
          $ref: "#/components/responses/Error"
      summary: Delete a Record
      tags:
      - metadata
      x-accepts:
      - application/json
      x-tags:
      - tag: metadata
    get:
      description: "Retrieves a specific metadata record value, or optionally the\
        \ full record view."
      parameters:
      - description: "The domain of the metadata record, representing which platform\
          \ or third party it relates to."
        explode: false
        in: path
        name: domain
        required: true
        schema:
          example: JANUS
          type: string
        style: simple
      - description: The registered name of the metadata record.
        explode: false
        in: path
        name: name
        required: true
        schema:
          example: overview
          type: string
        style: simple
      - description: Whether to return the full entity object instead of just the
          parsed JSON value.
        explode: true
        in: query
        name: full
        required: false
        schema:
          default: false
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_meta__domain___name__get_200_response"
          description: "Either a full metadata object, or the value of the record\
            \ parsed into JSON."
        "500":
          $ref: "#/components/responses/Error"
      summary: Retrieve a Record
      tags:
      - metadata
      x-accepts:
      - application/json
      x-tags:
      - tag: metadata
    post:
      description: Creates or updates a metadata record with a new value object.
      parameters:
      - description: "The domain of the metadata record, representing which platform\
          \ or third party it relates to."
        explode: false
        in: path
        name: domain
        required: true
        schema:
          example: JANUS
          type: string
        style: simple
      - description: The registered name of the metadata record.
        explode: false
        in: path
        name: name
        required: true
        schema:
          example: overview
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_meta__domain___name__post_request"
        description: The new JSON value for the record.
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SENSE_Metadata"
          description: A full metadata object.
        "500":
          $ref: "#/components/responses/Error"
      summary: Register a Record
      tags:
      - metadata
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: metadata
  /meta/{domain}/{name}/update:
    put:
      description: Updates a metadata record with updates to specific values.
      parameters:
      - description: "The domain of the metadata record, representing which platform\
          \ or third party it relates to."
        explode: false
        in: path
        name: domain
        required: true
        schema:
          example: JANUS
          type: string
        style: simple
      - description: The registered name of the metadata record.
        explode: false
        in: path
        name: name
        required: true
        schema:
          example: overview
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: "#/components/schemas/SENSE_Metadata_Update_DTO"
              type: array
        description: A JSON array of update objects.
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/paths/~1meta~1%7Bdomain%7D~1%7Bname%7D/post/responses/200/content/application~1json/schema"
          description: A full metadata object.
        "500":
          $ref: "#/components/responses/Error"
      summary: Updates a Record
      tags:
      - metadata
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: metadata
  /meta/{domain}/{name}/policy:
    get:
      description: Gets the list of associated policies with a metadata record.
      parameters:
      - description: "The domain of the metadata record, representing which platform\
          \ or third party it relates to."
        explode: false
        in: path
        name: domain
        required: true
        schema:
          example: JANUS
          type: string
        style: simple
      - description: The registered name of the metadata record.
        explode: false
        in: path
        name: name
        required: true
        schema:
          example: overview
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/paths/~1meta~1%7Bdomain%7D~1%7Bname%7D~1policy/post/responses/200/content/application~1json/schema"
                type: array
          description: An array of policy objects.
        "500":
          $ref: "#/components/responses/Error"
      summary: Retrieve Record Policies
      tags:
      - metadata
      x-accepts:
      - application/json
      x-tags:
      - tag: metadata
    post:
      description: Creates or updates a metadata policy against the specified record.
      parameters:
      - description: "The domain of the metadata record, representing which platform\
          \ or third party it relates to."
        explode: false
        in: path
        name: domain
        required: true
        schema:
          example: JANUS
          type: string
        style: simple
      - description: The registered name of the metadata record.
        explode: false
        in: path
        name: name
        required: true
        schema:
          example: overview
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SENSE_Metadata_Policy"
        description: A policy update request object.
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SENSE_Metadata_Policy_1"
          description: A single metadata policy object.
        "500":
          $ref: "#/components/responses/Error"
      summary: Register a Policy
      tags:
      - metadata
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: metadata
  /meta/{domain}/{name}/policy/id/{id}:
    get:
      description: Gets the a specific policy by its ID.
      parameters:
      - description: "The domain of the metadata record, representing which platform\
          \ or third party it relates to."
        explode: false
        in: path
        name: domain
        required: true
        schema:
          example: JANUS
          type: string
        style: simple
      - description: The registered name of the metadata record.
        explode: false
        in: path
        name: name
        required: true
        schema:
          example: overview
          type: string
        style: simple
      - description: The specified ID number of the policy.
        explode: false
        in: path
        name: id
        required: true
        schema:
          example: 591
          type: number
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/paths/~1meta~1%7Bdomain%7D~1%7Bname%7D~1policy/post/responses/200/content/application~1json/schema"
          description: A single metadata policy object.
        "500":
          $ref: "#/components/responses/Error"
      summary: Retrieve a Policy
      tags:
      - metadata
      x-accepts:
      - application/json
      x-tags:
      - tag: metadata
...
components:
  responses:
    Error:
      description: Internal server fault or standardized error response.
...
  schemas:
...
    _meta__domain___name__get_200_response:
      oneOf:
      - $ref: "#/paths/~1meta~1%7Bdomain%7D~1%7Bname%7D/post/responses/200/content/application~1json/schema"
      - type: object
    _meta__domain___name__post_request:
      oneOf:
      - type: object
      - type: string
    SENSE_Metadata:
      description: This schema represents a record of the SENSE metadata service.
      example:
        owner: ajimenez
        data: |-
          {
            "controller": {
              "foo": "bar"
            },
          ...
        edited: 2023-07-11 12:14:09.0
        created: 2023-07-03 10:39:10.0
        name: JANUS/overview
      properties:
        name:
          description: The unique compsite id of the metadata record. Composed of
            a service domain and record name separated by a slash.
          example: JANUS/overview
          type: string
        data:
          description: "The value of the record, encoded as a JSON in string format."
          example: |-
            {
              "controller": {
                "foo": "bar"
              },
            ...
          type: string
        owner:
          description: The user that created and is responsible for this record.
          example: ajimenez
          type: string
        created:
          description: The timestamp of the record's creation.
          example: 2023-07-03 10:39:10.0
          type: string
        edited:
          description: The timestamp of the record's last edit.
          example: 2023-07-11 12:14:09.0
          type: string
      required:
      - data
      - name
      - owner
      title: SENSE Metadata
      type: object
    SENSE_Metadata_Update_DTO:
      description: This schema represents an update object for SENSE metadata values.
      properties:
        path:
          description: "A JSON path for the specified value entry. Requires an existing\
            \ key at this path. Can use dot or bracket notation, as specified at https://github.com/json-path/JsonPath"
          example: $.agents.test_agent.host.fqdn
          type: string
        value:
          description: "The new value of the record, encoded as a JSON in string format.\
            \ Can set nested objects in this way."
          example: "[\"localhost\",\"{\\\"foo\\\":\\\"bar\\\",\\\"test\\\":[\\\"1\\\
            \", \\\"2\\\", \\\"3\\\"]}\"]"
          type: string
      required:
      - path
      - value
      title: SENSE Metadata Update DTO
      type: object
    SENSE_Metadata_Policy:
      description: This schema represents an access or validation policy for a metadata
        record.
      properties:
        name:
          description: The human-readable name of the policy.
          example: janus-agent-access
          type: string
        type:
          description: Whether the policy is involved with access or value validation.
          enum:
          - ACCESS
          - VALIDATION
          type: string
        principal:
          description: The username or group being moderated.
          example: playwright-tester
          type: string
        object:
          description: "The target of the policy, usually a metadata JSONPath."
          example: "$.agents[*].host"
          type: string
        level:
          description: The level at which the principal is given read or write privileges.
          enum:
          - NONE
          - READ
          - WRITE
          type: string
      title: SENSE Metadata Policy
      type: object
    SENSE_Metadata_Policy_1:
      description: This schema represents an access or validation policy for a metadata
        record.
      example:
        principal: playwright-tester
        level: NONE
        name: janus-agent-access
        id: 2131
        type: ACCESS
        object: "$.agents[*].host"
      properties:
        id:
          description: The policy id.
          example: 2131
          type: number
        name:
          description: The human-readable name of the policy.
          example: janus-agent-access
          type: string
        type:
          description: Whether the policy is involved with access or value validation.
          enum:
          - ACCESS
          - VALIDATION
          type: string
        principal:
          description: The username or group being moderated.
          example: playwright-tester
          type: string
        object:
          description: "The target of the policy, usually a metadata JSONPath."
          example: "$.agents[*].host"
          type: string
        level:
          description: The level at which the principal is given read or write privileges.
          enum:
          - NONE
          - READ
          - WRITE
          type: string
      required:
      - level
      - object
      - principal
      - type
      title: SENSE Metadata Policy
      type: object
...
  securitySchemes:
    OAuth2:
      flows:
        authorizationCode:
          authorizationUrl: https://auth.dev3.virnao.com/auth/realms/StackV/protocol/openid-connect/auth
          scopes:
            openid: OIDC Mode.
          tokenUrl: https://auth.dev3.virnao.com/auth/realms/StackV/protocol/openid-connect/token
      type: oauth2